Add files via upload
README.txt - readme file has detailed detailed information like usage, prerequisites, script file descriptions, installation, execution steps for ansible scripts to do network interfaces ping tests.

hosts-file - this is the file in which we define the IPs to which ansible should connect and do the ping test on all network interfaces. If user/password is specified for any listed IPs then ansible connects using the user/password defined under all:vars. If user/password is specified next to the IP then ansible connects using that user/password. Replace the username and password in this file to the username and password you want ansible to connect to the IPs listed. Replace the 10.x IP's with the range of IP's which you want ansible to connect to do ping tests.

ping-all-net-interfaces.yml - This is the file which ansible checks to see which network interfaces of the hosts (defined in hosts-file)  ping test has to be done. In this file define first 3 octets of the network interfaces (under item.startswith line) which you want to ping of hosts (which are defined in hosts-file). Change the 10.x, 192.x, 172.x IP subnet first 3 octets in this file to the octets corresponding to your hosts which you want to ping.